+++ eslint.config.mjs
@@ -20,6 +20,7 @@ const eslintConfig = [
       ],
       "react/no-unescaped-entities": "off",
       "@next/next/no-img-element": "off",
+      "@next/next/no-html-link-for-pages": "off",
     },
   },
   {